+++ src/main/java/com/barlinc/unusual_prehistory/datagen/UP2ItemTagProvider.java
@@ -398,6 +398,10 @@ public class UP2ItemTagProvider extends ItemTagsProvider {
                 TARIFYING_DISC.get()
         );
 
+        this.tag(ItemTags.LECTERN_BOOKS).add(
+                PALEOPEDIA.get()
+        );
+
         // Forge
         this.tag(UP2ItemTags.FRUITS).addTag(UP2ItemTags.FRUITS_GINKGO);
         this.tag(UP2ItemTags.FRUITS_GINKGO).add(GINKGO_FRUIT.get());