+++ src/main/java/com/barlinc/unusual_prehistory/items/PaleopediaItem.java
@@ -1,15 +1,46 @@
 package com.barlinc.unusual_prehistory.items;
 
-// todo: figure out how to get this working
-//public class PaleopediaItem extends ItemModBook {
-//
-//    public PaleopediaItem() {
-//        super();
-//    }
-//
-//    @Override
-//    public void appendHoverText(@NotNull ItemStack stack, @Nullable Level world, @NotNull List<Component> tooltip, @NotNull TooltipFlag flag) {
-//        tooltip.add(Component.translatable("item.unusual_prehistory.paleopedia.desc").withStyle(ChatFormatting.GRAY));
-//        super.appendHoverText(stack, world, tooltip, flag);
-//    }
-//}
+import com.barlinc.unusual_prehistory.registry.UP2Items;
+import net.minecraft.ChatFormatting;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.level.Level;
+import org.jetbrains.annotations.NotNull;
+import vazkii.patchouli.api.PatchouliAPI;
+
+import java.util.List;
+
+public class PaleopediaItem extends Item {
+
+    public PaleopediaItem(Properties properties) {
+        super(properties);
+    }
+
+    public static boolean isOpen() {
+        return BuiltInRegistries.ITEM.getKey(UP2Items.PALEOPEDIA.get()).equals(PatchouliAPI.get().getOpenBookGui());
+    }
+
+    @Override
+    public void appendHoverText(@NotNull ItemStack stack, Level level, List<Component> tooltip, @NotNull TooltipFlag flag) {
+        tooltip.add(Component.translatable("item.unusual_prehistory.paleopedia.desc").withStyle(ChatFormatting.GRAY));
+    }
+
+    @NotNull
+    @Override
+    public InteractionResultHolder<ItemStack> use(@NotNull Level level, Player player, @NotNull InteractionHand hand) {
+        ItemStack stack = player.getItemInHand(hand);
+        if (player instanceof ServerPlayer serverPlayer) {
+            PatchouliAPI.get().openBookGUI(serverPlayer, BuiltInRegistries.ITEM.getKey(this));
+            player.playSound(SoundEvents.BOOK_PAGE_TURN, 1F, 0.9F + level.getRandom ().nextFloat() * 0.25F);
+        }
+        return InteractionResultHolder.sidedSuccess(stack, level.isClientSide());
+    }
+}